Rename validation utility file shared by transaction and delivery
