+++ public/js/transaction-delivery-validate-util.js
